Fixed #9: generateUniqueIRI didn't generate a unique IRI.
The issue is addressed by adding a method `putAndReturnIndex` to `MapState` which remembers all values associated once
to a given key (or here: all watchedPropertyValues for a given template). If the given value is not associated yet,
the number of previously associated values is returned. If the given value was already associated, an empty value
is retured. That way `generateUniqueIRI` can generate a unique IRI for each new value in one step by appending the
returned number.